crossdock: Move yarpc server into its own package, making way for pure tch server
